Reestructurar componente Carrusel, agregar nuevo componente de inicio de sesión y definir fuente personalizada Copperplate
