+++ GYMPROMANAGER-APP/tailwind.config.js
@@ -5,6 +5,9 @@ module.exports = {
   presets: [require('nativewind/preset')],
   theme: {
     extend: {},
+    fontFamily: {
+      Copperplate: ['"Copperplate"', 'sans-serif'],
+    },
   },
   plugins: [],
 }
